+++ fractal/components/containers/LayerContainer/LayerContainer.jsx
@@ -2,20 +2,17 @@ import React from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
-const LayerContainer = ({ className, children, ...rest }) => {
+const LayerContainer = ({ className, ...rest }) => {
     const componentClass = 'c-layer-container';
     const gridClass = 'l-grid-wrapper l-grid-container-fluid';
     const componentClasses = classNames(componentClass, gridClass, className);
     return (
-        <div className={componentClasses} {...rest}>
-            {children}
-        </div>
+        <div className={componentClasses} {...rest}/>
     );
 };
 
 LayerContainer.propTypes = {
-    className: PropTypes.string,
-    children: PropTypes.node
+    className: PropTypes.string
 };
 
 LayerContainer.defaultProps = {};