+++ fractal/components/containers/PageLayer/PageLayer.jsx
@@ -4,19 +4,20 @@ import PropTypes from 'prop-types';
 
 import LayerContainer from '../LayerContainer/LayerContainer.jsx';
 
-const PageLayer = ({ className, children, constrained = true, ...rest }) => {
+const PageLayer = ({ className, containerClassName, children, constrained = true, ...rest }) => {
 
     const componentClass = 'c-page-layer';
+    const containerClass = `${componentClass}__container`;
     const constrainedClass = 'is-constrained';
-    const componentClasses = classNames(componentClass, className,
-        {[constrainedClass]: constrained});
+    const componentClasses = classNames(componentClass, className, {[constrainedClass]: constrained});
+    const containerClasses = classNames(containerClass, {[containerClassName]: containerClassName});
 
     return (
         <div className={componentClasses} {...rest}>
             {
                 constrained
                     ?
-                <LayerContainer>{children}</LayerContainer>
+                <LayerContainer className={containerClasses}>{children}</LayerContainer>
                     :
                 {children}
             }
@@ -26,6 +27,7 @@ const PageLayer = ({ className, children, constrained = true, ...rest }) => {
 
 PageLayer.propTypes = {
     className: PropTypes.string,
+    containerClassName: PropTypes.string,
     constrained: PropTypes.bool,
     children: PropTypes.node
 };